* From now on, compiled ASF can deal with annotations!    - ATisEqual is replaced by ATisEqualModuloAnnotations    - innermost and muasf2pt preserve annotations    - asfc has a '-a' option to activate it.    - gcc must be called with -DASC_WITH_ANNOS to select the correct      equality implementation.    - before calling asc_support_main, you must call      setKeepAnnotations(ATtrue) to make innermost and muasf2pt behave      properly.    - TODO: refactor asc_support_main to have an extra argument.      This is not doable for release 1.5

+++ asc-support/lib/asc-main.c
@@ -209,7 +209,7 @@ int asc_support_main(ATerm *bottomOfStack, int argc, char *argv[],
 	char buf[BUFSIZ];
 	int size, resident, shared, trs, lrs, drs, dt;
 
-	sprintf(buf, "/proc/%d/statm", getpid());
+	sprintf(buf, "/proc/%d/statm", (int) getpid());
 	file = fopen(buf, "r");
 	if (file) {
 	  fscanf(file, "%d %d %d %d %d %d %d",
